requirements.txt
Flask==3.1.0
discord.py==2.5.2
flask-cors==5.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.5
stripe==12.4.0
